Added support for loading external fonts. PDFStylesMSSAdapter now looks in the pdf/ttf section of the mss file to find a font definition to load if the specified font for a section is not one of the standard fonts.
There is still a catch however. Relative paths to font files differs depending on if I build from top or from a module. I need to create a GenaratorContext containing to begin with the rootDir that each generator receives on start, just to solve this problem. Then the GeneratorContext gets passed around, and any more information needed can just be added to the context.