mdd-4
Added <div class="name">...</div> support. The markdown parser will recognize this construct and add a model for it. HTML and markdown generators simply reproduce the tag. The PDF generator however keeps track of the current div if any and any parent divs and applies them when using MSS to format result.