mdd-20 The end div gets rendered but without <> in PDFGenerator!
This turned out to be correct since I did not put an empty line before the end div in my md document. A paragraph is ended by an empty line. Without an empty line before the end div then it becomes part of the paragraph. The start div however is only one line so no empty line between start div and paragraph is OK.